#250 use filter for corpus creation

+++ core-initcorpus/src/main/java/cc/topicexplorer/initcorpus/RemoveDuplicates.java
@@ -1,6 +1,7 @@
 package cc.topicexplorer.initcorpus;
 
 import java.sql.Connection;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Properties;
@@ -13,7 +14,6 @@ import cc.commandmanager.core.Context;
 import cc.commandmanager.core.ResultState;
 
 import com.google.common.collect.Sets;
-
 
 public class RemoveDuplicates implements Command {
 
@@ -29,28 +29,49 @@ public class RemoveDuplicates implements Command {
 
 		this.tableName = dbName + ".DUPLICATES";
 
-
+		String topicExplorer_DefinitionIsPresent = "";
 
 		try {
-			Statement stmt = crawlManagerConnection.createStatement();
-
-			stmt.executeUpdate("create table " + this.tableName + " as "
-					+ "select URL_MD5, min(DOCUMENT_DATE) mindate, min(DOCUMENT_ID) minid from " + dbName + ".orgTable_meta "
-					+ "group by URL_MD5 having count(*) > 1");
-
-			stmt.executeUpdate("create index url_idx on " + this.tableName + "(URL_MD5,mindate,minid)");
-
-			stmt.executeUpdate("delete " + dbName + ".orgTable_meta, " + dbName + ".orgTable_text from "
-					+ dbName + ".orgTable_meta, " + dbName + ".orgTable_text, " + this.tableName
-					+ " where orgTable_meta.DOCUMENT_ID=orgTable_text.DOCUMENT_ID and orgTable_meta.URL_MD5=" + this.tableName + ".URL_MD5 and "
-					+ "(orgTable_meta.DOCUMENT_DATE>" + this.tableName + ".mindate or (orgTable_meta.DOCUMENT_DATE=" + this.tableName + ".mindate "
-					+ "and orgTable_meta.DOCUMENT_ID>" + this.tableName + ".minid))");
-
-			stmt.executeUpdate("drop table " + this.tableName);
-
+			Statement checkStmt = crawlManagerConnection.createStatement();
+			ResultSet resultSetOfCheck = checkStmt.executeQuery("select count(*) as IS_THERE "
+					+ "from  INFORMATION_SCHEMA.Tables "
+					+ "where TABLE_NAME='TOPIC_EXPLORER_DEFINITION' and TABLE_SCHEMA=DATABASE();");
+			if (resultSetOfCheck.next()) {
+				topicExplorer_DefinitionIsPresent = resultSetOfCheck.getString("IS_THERE");
+			}
 		} catch (SQLException e) {
-			logger.error("Table " + this.tableName + " could not be filled properly.");
-			return ResultState.failure("Table " + this.tableName + " could not be filled properly.", e);
+			logger.error("Can not check for table TOPIC_EXPLORER_DEFINITION in mangagement database.");
+			return ResultState.failure(
+					"Can not check for table TOPIC_EXPLORER_DEFINITION in mangagement database. Table "
+							+ this.tableName + " could not be filled properly.", e);
+		}
+
+		if ("0".equals(topicExplorer_DefinitionIsPresent)) {
+			// Old Crawl Management Database, this may contain duplicates in the
+			// documents
+			// that need to be handled.
+			try {
+				Statement stmt = crawlManagerConnection.createStatement();
+
+				stmt.executeUpdate("create table " + this.tableName + " as "
+						+ "select URL_MD5, min(DOCUMENT_DATE) mindate, min(DOCUMENT_ID) minid from " + dbName
+						+ ".orgTable_meta " + "group by URL_MD5 having count(*) > 1");
+
+				stmt.executeUpdate("create index url_idx on " + this.tableName + "(URL_MD5,mindate,minid)");
+
+				stmt.executeUpdate("delete " + dbName + ".orgTable_meta, " + dbName + ".orgTable_text from " + dbName
+						+ ".orgTable_meta, " + dbName + ".orgTable_text, " + this.tableName
+						+ " where orgTable_meta.DOCUMENT_ID=orgTable_text.DOCUMENT_ID and orgTable_meta.URL_MD5="
+						+ this.tableName + ".URL_MD5 and " + "(orgTable_meta.DOCUMENT_DATE>" + this.tableName
+						+ ".mindate or (orgTable_meta.DOCUMENT_DATE=" + this.tableName + ".mindate "
+						+ "and orgTable_meta.DOCUMENT_ID>" + this.tableName + ".minid))");
+
+				stmt.executeUpdate("drop table " + this.tableName);
+
+			} catch (SQLException e) {
+				logger.error("Table " + this.tableName + " could not be filled properly.");
+				return ResultState.failure("Table " + this.tableName + " could not be filled properly.", e);
+			}
 		}
 		return ResultState.success();
 
@@ -75,7 +96,5 @@ public class RemoveDuplicates implements Command {
 	public Set<String> getOptionalBeforeDependencies() {
 		return Sets.newHashSet();
 	}
-
-
 
 }
